refactor(#40): ContentInput 로직 리팩토링
- atom으로 이동
