+++ src/components/_common/atoms/ContentInput/style.ts
