+++ src/screens/FeedCreateScreen/index.tsx
@@ -1,11 +1,12 @@
 import ImageUploader from '@components/ImageUploader';
 import S from './style';
 import { useState } from 'react';
-import ContentInput from '@components/ContentInput';
 import Button from '@components/_common/atoms/Button';
+import ContentInput from '@components/_common/atoms/ContentInput';
 
 export default function FeedCreateScreen() {
   const [images, setImages] = useState<string[]>([]);
+  const [content, setContent] = useState<string>('');
 
   const handlePress = () => {};
 
@@ -18,11 +19,13 @@ export default function FeedCreateScreen() {
         </S.ImageForm>
         <S.ContentWrapper>
           <S.ContentLabel>본문</S.ContentLabel>
-          <ContentInput />
+          <ContentInput text={content} onChange={setContent} />
         </S.ContentWrapper>
       </S.FeedCreateContainer>
       <S.ButtonWrapper>
-        <Button onPress={handlePress}>완료</Button>
+        <Button size='fillContainer' color='primary' shape='round' onPress={handlePress}>
+          완료
+        </Button>
       </S.ButtonWrapper>
     </S.Layout>
   );
